requirements.txt
DataSpyre==0.2.0
matplotlib==1.4.3
seaborn==0.6.0
nltk==3.0.4
pandas==0.16.2
pymongo==3.0.3
